Added structure, some graphical changes
