+++ lib/main.dart
@@ -1,21 +1,22 @@
-import 'package:flu_new/profile_page.dart';
+import 'package:flu_new/view/profile_page.dart';
 import 'package:flutter/material.dart';
-import 'home_page.dart';
-// import 'search_page.dart';
+import 'view/home_page.dart';
+
 void main() {
-  runApp(const MyApp());
+  runApp(const App());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+class App extends StatelessWidget {
+  const App({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: '',
       theme: ThemeData(
-        // colorScheme: ColorScheme.fromSeed(seedColor: const Color.fromARGB(255, 107, 215, 53)),
-        colorScheme: ColorScheme.fromSeed(seedColor: const Color.fromARGB(255, 238, 240, 238)),
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: const Color.fromARGB(255, 238, 240, 238),
+        ),
         useMaterial3: true,
       ),
       home: const HomeScreen(),
@@ -38,22 +39,7 @@ class HomeScreen extends StatelessWidget {
             UserProfileScreen(),
           ],
         ),
-        bottomNavigationBar: Material(
-          // color: Theme.of(context).colorScheme.primary,
-          color: Colors.white,
-          child: const TabBar(
-            tabs: [
-              Tab(icon: Icon(Icons.list), text: 'Listing'),
-              // Tab(icon: Icon(Icons.add), text: 'Add'),
-              Tab(icon: Icon(Icons.person), text: 'Profile')
-            ],
-            labelColor: Colors.green,
-            unselectedLabelColor: Colors.lightGreen,
-            indicatorColor: Colors.green,
-          ),
-        ),
       ),
     );
   }
 }
-
